Dashboard page code refactor
