+++ src/components/Sidebar.tsx
@@ -1,5 +1,5 @@
 import { Avatar } from "@fluentui/react-components";
-import { sideBarItems } from "../constants";
+import { Messages, sideBarItems } from "../constants";
 
 interface ISidebarProps {
   selectedItem: string;
@@ -8,13 +8,13 @@ interface ISidebarProps {
 
 const Sidebar = ({ selectedItem, handleItemClick }: ISidebarProps) => {
   return (
-    <div className="flex flex-row gap-8 w-1/5">
+    <div className="flex flex-row gap-8 w-1/6">
       <section className="flex flex-col gap-y-6 shrink-0 items-center">
         <img src="/vertx-icon.png" alt="vertx" className="w-12 h-12" />
         <Avatar shape="circular" className="bg-blue-500" initials={"SP"} />
       </section>
       <section className="flex flex-col justify-center mb-auto">
-        <h1 className="mb-4 text-xl font-semibold">Vertx Labs, Inc</h1>
+        <h1 className="mb-4 text-xl font-semibold">{Messages.vertxLabs}</h1>
         {sideBarItems.map((item: string) => (
           <div
             key={item}
