--- src/containers/analytics/index.ts
@@ -1,3 +1,0 @@
-import AnalyticsContainer from "./AnalyticsContainer";
-
-export { AnalyticsContainer };
