+++ src/containers/index.ts
@@ -1,10 +1,10 @@
 import SidebarContainer from "./SidebarContainer";
-import { AnalyticsContainer } from "./analytics";
+import AnalyticsContainer from "./AnalyticsContainer";
 import OverviewContainer from "./OverviewContainer";
 import DemographicsContainer from "./DemographicsContainer";
 export {
   SidebarContainer,
-  AnalyticsContainer,
   OverviewContainer,
   DemographicsContainer,
+  AnalyticsContainer,
 };
